god ducking dammit.... Day 17
You know... It would've been very easy if I hadn't used vecdeques to store to_fall and to_spread
at first, because after some iterations they produce a looooot of duplicates which slow the entire thing down. So after about 60% of the way everything just GRIIIINDS to a halt because the to_fall vector contains about 60 ducking thousand elements!!! (99% of which are duplicates mind you).

Yeah so after hours of debugging I finally noticed and switched to HashSets.... and it runs. Go figure!